pipeline_module/requirements.txt
# Fairness Pipeline Development Toolkit - Requirements
# 48-hour academic demo scope

# ============================================================================
# CORE DEPENDENCIES (Must Have)
# ============================================================================

# Fairness Libraries (unified interface)
fairlearn>=0.9.0
aif360>=0.5.0

# ML Frameworks
scikit-learn>=1.3,<2.0
xgboost>=2.0.0
torch>=2.0.0

# Data & Statistics
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# MLOps Integration
mlflow>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# ============================================================================
# OPTIONAL DEPENDENCIES (Nice to Have)
# ============================================================================

# Advanced Statistics (if time permits for monitoring module)
statsmodels>=0.14.0

# Notebook Support
jupyter>=1.0.0
ipywidgets>=8.0.0

# ============================================================================
# DEVELOPMENT DEPENDENCIES (For 48-hour sprint)
# ============================================================================

# Code Quality (minimal - focus on functionality)
black>=23.0.0
flake8>=6.0.0


# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration
pyyaml>=5.4.0

# Deep Learning (optional - only if using training_module with neural networks)
torch>=1.9.0

# MLflow (optional - for experiment tracking)
mlflow>=1.20.0

# Jupyter (optional - for notebook conversion)
jupytext>=1.11.0
jupyter>=1.0.0

# Testing
pytest>=6.2.0

# ============================================================================
# NOTES
# ============================================================================
# - Pin major versions only for stability
# - Avoid heavy dependencies (TensorFlow, etc.) to speed up installation
# - Focus on sklearn + Fairlearn for 48-hour scope
# - PyTorch included but may be deprioritized if time constrained
# - All packages tested on Python 3.9+